+++ src/main/java/com/foolsix/fancyenchantments/effect/CritRateBoost.java
@@ -1,7 +1,7 @@
 package com.foolsix.fancyenchantments.effect;
 
+import com.foolsix.fancyenchantments.enchantment.util.EnchUtils;
 import net.minecraft.core.particles.ParticleTypes;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
 import net.minecraft.world.effect.MobEffectInstance;
@@ -11,6 +11,7 @@ import net.minecraftforge.event.entity.living.LivingHurtEvent;
 
 public class CritRateBoost extends MobEffect {
     public static final String CRIT_RATE_BOOST_NAME = "crit_rate_boost";
+
     CritRateBoost() {
         super(MobEffectCategory.BENEFICIAL, 0xFF0000);
     }
@@ -24,7 +25,7 @@ public class CritRateBoost extends MobEffect {
                 int amplifier = effectInstance.getAmplifier();
                 if (Math.random() < 0.2 * amplifier) {
                     e.setAmount(e.getAmount() * 1.5f);
-                    ((ServerLevel) attacker.level).sendParticles(ParticleTypes.CRIT, victim.getX(), victim.getY(), victim.getZ(), 10, 0.2D, 0.7D, 0.1D, 0);
+                    EnchUtils.generateSimpleParticleAroundEntity(victim, ParticleTypes.CRIT);
                 }
             }
         }