+++ src/main/java/com/foolsix/fancyenchantments/enchantment/UnyieldingSpirit.java
@@ -3,9 +3,9 @@ package com.foolsix.fancyenchantments.enchantment;
 import com.foolsix.fancyenchantments.FancyEnchantments;
 import com.foolsix.fancyenchantments.capability.TimeToLiveHelper;
 import com.foolsix.fancyenchantments.enchantment.EssentiaEnch.HolyEnchantment;
+import com.foolsix.fancyenchantments.enchantment.util.EnchUtils;
 import com.foolsix.fancyenchantments.util.ModConfig;
 import net.minecraft.core.particles.ParticleTypes;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
@@ -17,9 +17,9 @@ import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraftforge.event.entity.living.LivingDamageEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
 import net.minecraftforge.event.entity.living.LivingEquipmentChangeEvent;
-import net.minecraftforge.event.entity.living.LivingHurtEvent;
 
 
 public class UnyieldingSpirit extends HolyEnchantment {
@@ -41,7 +41,7 @@ public class UnyieldingSpirit extends HolyEnchantment {
         return 75;
     }
 
-    public void preventDeath(LivingHurtEvent e) {
+    public void preventDeath(LivingDamageEvent e) {
         if (e.getEntity() instanceof ServerPlayer player && EnchantmentHelper.getEnchantmentLevel(this, player) > 0) {
             if (e.getSource() != DamageSource.OUT_OF_WORLD && e.getAmount() >= player.getHealth()) {
                 if (TimeToLiveHelper.getTtl(player) == -1) {
@@ -67,8 +67,8 @@ public class UnyieldingSpirit extends HolyEnchantment {
             if (TimeToLiveHelper.getTtl(player) != -1) {
                 TimeToLiveHelper.setTtl(player, -1);
                 player.setHealth(player.getMaxHealth());
-                ((ServerLevel) player.level).sendParticles(ParticleTypes.TOTEM_OF_UNDYING, player.getX(), player.getY(), player.getZ(), 100, 1.0D, 1.0D, 1.0D, 0.5);
-                player.level.playSound(player, player.blockPosition(), SoundEvents.TOTEM_USE, SoundSource.PLAYERS, 4.0F, 1.5F);
+                EnchUtils.generateSimpleParticleAroundEntity(player, ParticleTypes.TOTEM_OF_UNDYING, 100, 1.0, 1.0, 1.0, 0.5);
+                player.getLevel().playSound(null, player.blockPosition(), SoundEvents.TOTEM_USE, SoundSource.PLAYERS, 3.0F, 1.0F);
             }
         }
     }